requirements.txt
aniso8601==9.0.1
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-RESTful==0.3.9
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scikit-surprise==1.1.1
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
Werkzeug==2.1.1
zipp==3.8.0
